#29 리팩토링
