+++ src/components/Play/index.js
@@ -1,5 +1,6 @@
 import Pedigree from "./Pedigree";
 import Total from "./Total";
 import Dice from "./Dice";
+import PedigreeInfo from "./PedigreeInfo";
 
-export { Pedigree, Total, Dice };
+export { Pedigree, Total, Dice, PedigreeInfo };
