+++ src/screens/FriendsScreen/FriendsScreenContainer.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import FriendsScreenPresenter from "./FriendsScreenPresenter";
+import { dummy } from "../../dummies/friends";
 
-import dummy from "./dummy1";
 export default function () {
   const [tabNumber, setTabNumber] = useState(0);
   const [friendsList, setFriendsList] = useState(null);
@@ -10,16 +10,20 @@ export default function () {
   const onClickLeftTab = () => {
     setTabNumber(0);
   };
+
   const onClickRightTab = () => {
     setTabNumber(1);
   };
+
   const saveDatas = () => {
     setFriendsList(dummy);
     setFriendsRequestList(dummy);
   };
+
   useEffect(() => {
     saveDatas();
   }, []);
+
   return (
     <FriendsScreenPresenter
       tabNumber={tabNumber}
