+++ src/screens/FriendsScreen/index.js
@@ -1,4 +1,3 @@
-// import FriendScreen from "./FriendsScreen";
 import FriendsScreenContainer from "./FriendsScreenContainer";
 
 export default FriendsScreenContainer;
